+++ src/main/java/com/gulbi/Backend/domain/rental/product/service/product/ProductServiceImpl.java
@@ -3,7 +3,9 @@ package com.gulbi.Backend.domain.rental.product.service.product;
 import com.gulbi.Backend.domain.rental.product.dto.product.ProductOverViewResponse;
 import com.gulbi.Backend.domain.rental.product.dto.product.request.*;
 import com.gulbi.Backend.domain.rental.product.dto.product.response.ProductDetailResponseDto;
-import com.gulbi.Backend.domain.rental.product.vo.image.ProductImageCollection;
+import com.gulbi.Backend.domain.rental.product.service.product.register.ProductRegistrationService;
+import com.gulbi.Backend.domain.rental.product.service.product.search.ProductSearchService;
+import com.gulbi.Backend.domain.rental.product.service.product.update.ProductUpdatingService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
